--- src/main/java/com/powsybl/gd/model/BusNode.java
@@ -1,4 +1,0 @@
-package com.powsybl.gd.model;
-
-public class BusNode {
-}
