--- src/main/java/com/powsybl/gd/model/Node.java
@@ -1,19 +1,0 @@
-/**
- * Copyright (c) 2021, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package com.powsybl.gd.model;
-
-import java.util.List;
-
-public interface Node {
-
-    String getDiagramId();
-
-    List<? extends Edge> getAdjacentEdges();
-
-    void setPosition(double x, double y);
-
-}
