+++ src/main/java/com/powsybl/nad/model/BranchEdge.java
@@ -4,14 +4,12 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.gd.model;
+package com.powsybl.nad.model;
 
 import java.util.List;
 
-public interface Edge {
+public interface BranchEdge extends Edge {
 
-    String getDiagramId();
-
-    List<? extends Node> getAdjacentNodes();
+    List<VoltageLevelNode> getAdjacentNodes();
 
 }
