+++ src/main/java/com/powsybl/nad/model/BusNode.java
@@ -1,0 +1,4 @@
+package com.powsybl.nad.model;
+
+public class BusNode {
+}
