+++ src/main/java/com/powsybl/nad/model/Edge.java
@@ -4,12 +4,14 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.gd.model;
+package com.powsybl.nad.model;
 
 import java.util.List;
 
-public interface BranchEdge extends Edge {
+public interface Edge {
 
-    List<VoltageLevelNode> getAdjacentNodes();
+    String getDiagramId();
+
+    List<? extends Node> getAdjacentNodes();
 
 }
