+++ src/main/java/com/powsybl/nad/model/Node.java
@@ -4,10 +4,16 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.gd.model;
+package com.powsybl.nad.model;
 
-public class TwoWindingsTransformerEdge extends AbstractTransformerEdge {
-    protected TwoWindingsTransformerEdge(String id, VoltageLevelNode node1, VoltageLevelNode node2) {
-        super(id, node1, node2);
-    }
+import java.util.List;
+
+public interface Node {
+
+    String getDiagramId();
+
+    List<? extends Edge> getAdjacentEdges();
+
+    void setPosition(double x, double y);
+
 }
