+++ src/main/java/com/powsybl/nad/model/ThreeWindingsTransformerEdge.java
@@ -1,0 +1,13 @@
+/**
+ * Copyright (c) 2021, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package com.powsybl.nad.model;
+
+public class ThreeWindingsTransformerEdge extends AbstractTransformerEdge {
+    protected ThreeWindingsTransformerEdge(String id, VoltageLevelNode node1, VoltageLevelNode node2, VoltageLevelNode node3) {
+        super(id, node1, node2, node3);
+    }
+}
